+++ packages/troika-core/CHANGELOG.md
@@ -3,6 +3,17 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.41.0](https://github.com/protectwise/troika/compare/v0.40.0...v0.41.0) (2021-04-19)
+
+
+### Bug Fixes
+
+* **troika-core:** set animation/transition first when using .update() ([8ebf59d](https://github.com/protectwise/troika/commit/8ebf59dc9e933400f8302c0e5d99d15411b248c0))
+
+
+
+
+
 # [0.40.0](https://github.com/protectwise/troika/compare/v0.39.2...v0.40.0) (2021-02-28)
 
 **Note:** Version bump only for package troika-core