+++ packages/troika-flex-layout/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.41.1](https://github.com/protectwise/troika/compare/v0.41.0...v0.41.1) (2021-04-26)
+
+**Note:** Version bump only for package troika-flex-layout
+
+
+
+
+
+# [0.41.0](https://github.com/protectwise/troika/compare/v0.40.0...v0.41.0) (2021-04-19)
+
+**Note:** Version bump only for package troika-flex-layout
+
+
+
+
+
 # [0.40.0](https://github.com/protectwise/troika/compare/v0.39.2...v0.40.0) (2021-02-28)
 
 **Note:** Version bump only for package troika-flex-layout